backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
playwright==1.40.0
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic[email]==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
aiofiles==23.2.1
httpx==0.25.2
redis==5.0.1
celery==5.3.4
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
asyncio-mqtt==0.16.1
pandas==2.1.4
openpyxl==3.1.2
